src/web/requirements.txt
openai==1.50.2
pydantic==2.9.2
python-dotenv==1.0.1
tiktoken==0.7.0
azure-search-documents==11.5.1
azure-identity==1.17.1
streamlit==1.38.0
langchain==0.3.1
langchain-core==0.3.6
langchain-community==0.3.1
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
langchain-qdrant==0.1.4
langgraph==0.2.29
langgraph-checkpoint==1.0.13
beautifulsoup4==4.12.3